Property

+++ src/main/java/com/project/PropertyVersatile/controller/PropertyController.java
@@ -80,8 +80,15 @@ public class PropertyController {
     @GetMapping("/{propertyId}/delete")
     public String deleteProperty(@PathVariable int propertyId, Model model) {
         try {
-            propertyService.deleteProperty(propertyId);
-            // Redirect to the "/properties" URL after successfully deleting the property
+            boolean hasMaintenanceRequests = propertyService.hasMaintenanceRequests(propertyId);
+
+            if (hasMaintenanceRequests) {
+                model.addAttribute("maintenanceRequestsPresent", true);
+            } else {
+                propertyService.deleteProperty(propertyId);
+                model.addAttribute("action", "delete");
+            }
+
             return "redirect:/properties";
         } catch (Exception e) {
             e.printStackTrace();
@@ -89,7 +96,6 @@ public class PropertyController {
             return "error";
         }
     }
-
     
     @PostMapping("/update")
     public String updateProperty(@ModelAttribute Property property, Model model) {
